+++ src/bindings/python/python.cc
@@ -14,7 +14,6 @@
 #include "midgard/util.h"
 #include "tyr/actor.h"
 
-
 namespace {
 
 // statically set the config file and configure logging, throw if you never configured
@@ -54,28 +53,39 @@ void py_configure(const std::string& config_file) {
 }
 } // namespace
 
-
 namespace py = pybind11;
 
-template <typename... Args>
-using overload_cast_ = pybind11::detail::overload_cast_impl<Args...>;
+template <typename... Args> using overload_cast_ = pybind11::detail::overload_cast_impl<Args...>;
 
-PYBIND11_MODULE(valhalla_python, m)
-{
-    m.def("Configure", py_configure);
+PYBIND11_MODULE(valhalla_python, m) {
+  m.def("Configure", py_configure);
 
-    py::class_<valhalla::tyr::actor_t, std::shared_ptr<valhalla::tyr::actor_t>>(m, "Actor")
-        .def(py::init<>([](){
-            return std::make_shared<valhalla::tyr::actor_t>(configure(), true);
-        }))
-        .def("Route", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::route), "Calculates a route.")
-        .def("Locate", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::locate), "Provides information about nodes and edges.")
-        .def("OptimizedRoute", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::optimized_route), "Optimizes the order of a set of waypoints by time.")
-        .def("Matrix", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::matrix), "Computes the time and distance between a set of locations and returns them as a matrix table.")
-        .def("Isochrone", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::isochrone), "Calculates isochrones and isodistances.")
-        .def("TraceRoute", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::trace_route), "Map-matching for a set of input locations, e.g. from a GPS.")
-        .def("TraceAttributes", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::trace_attributes), "Returns detailed attribution along each portion of a route calculate from a set of input locations, e.g. from a GPS trace.")
-        .def("Height", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::height), "Provides elevation data for a set of input geometries.")
-        .def("TransitAvailable", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::transit_available), "Lookup if transit stops are available in a defined radius around a set of input locations.")
-        .def("Expansion", overload_cast_<const  std::string &>()(&valhalla::tyr::actor_t::expansion), "Returns all road segments which were touched by the routing algorithm during the search.");
+  py::class_<valhalla::tyr::actor_t, std::shared_ptr<valhalla::tyr::actor_t>>(m, "Actor")
+      .def(py::init<>([]() { return std::make_shared<valhalla::tyr::actor_t>(configure(), true); }))
+      .def("Route", overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::route),
+           "Calculates a route.")
+      .def("Locate", overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::locate),
+           "Provides information about nodes and edges.")
+      .def("OptimizedRoute",
+           overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::optimized_route),
+           "Optimizes the order of a set of waypoints by time.")
+      .def(
+          "Matrix", overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::matrix),
+          "Computes the time and distance between a set of locations and returns them as a matrix table.")
+      .def("Isochrone", overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::isochrone),
+           "Calculates isochrones and isodistances.")
+      .def("TraceRoute", overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::trace_route),
+           "Map-matching for a set of input locations, e.g. from a GPS.")
+      .def(
+          "TraceAttributes",
+          overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::trace_attributes),
+          "Returns detailed attribution along each portion of a route calculate from a set of input locations, e.g. from a GPS trace.")
+      .def("Height", overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::height),
+           "Provides elevation data for a set of input geometries.")
+      .def(
+          "TransitAvailable",
+          overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::transit_available),
+          "Lookup if transit stops are available in a defined radius around a set of input locations.")
+      .def("Expansion", overload_cast_<const std::string&>()(&valhalla::tyr::actor_t::expansion),
+           "Returns all road segments which were touched by the routing algorithm during the search.");
 }