+++ src/tyr/actor.cc
@@ -81,7 +81,6 @@ std::string actor_t::route(const std::string& request_str) {
   return actor_t::route(request_str, NULL, NULL);
 }
 
-
 std::string
 actor_t::locate(const std::string& request_str, const std::function<void()>* interrupt, Api* api) {
   // set the interrupts